requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.22
httpx==0.28.1
openai==1.107.2
langgraph==0.6.6
langchain-core==0.3.75
lancedb==0.24.3
sentence-transformers==5.1.0
pandas==2.3.2
openpyxl==3.1.5
xlrd==2.0.1
pymupdf==1.26.4
Pillow==11.3.0
orjson==3.10.18
sqlalchemy==2.0.43
pytest==9.0.2
